添加亮度和对比度的实现

+++ UI/ColorVision.ImageEditor/ImageViewConfig.cs
@@ -92,14 +92,14 @@ namespace ColorVision.ImageEditor
         public event EventHandler BalanceChanged;
 
 
-        public float RedBalance { get => _RedBalance; set { _RedBalance = value; NotifyPropertyChanged(); BalanceChanged?.Invoke(this, new EventArgs()); } }
-        private float _RedBalance = 1;
+        public double RedBalance { get => _RedBalance; set { _RedBalance = value; NotifyPropertyChanged(); BalanceChanged?.Invoke(this, new EventArgs()); } }
+        private double _RedBalance = 1;
 
-        public float GreenBalance { get => _GreenBalance; set { _GreenBalance = value; NotifyPropertyChanged(); BalanceChanged?.Invoke(this, new EventArgs()); } }
-        private float _GreenBalance = 1;
+        public double GreenBalance { get => _GreenBalance; set { _GreenBalance = value; NotifyPropertyChanged(); BalanceChanged?.Invoke(this, new EventArgs()); } }
+        private double _GreenBalance = 1;
 
-        public float BlueBalance { get => _BlueBalance; set { _BlueBalance = value; NotifyPropertyChanged(); BalanceChanged?.Invoke(this, new EventArgs()); } }
-        private float _BlueBalance = 1;
+        public double BlueBalance { get => _BlueBalance; set { _BlueBalance = value; NotifyPropertyChanged(); BalanceChanged?.Invoke(this, new EventArgs()); } }
+        private double _BlueBalance = 1;
 
     }
 }
